This MR adds admin login and admin features to the app

See merge request cse1105/2022-2023/teams/oopp-team-62!84

+++ client/src/main/java/client/scenes/BoardOverviewController.java
@@ -1,6 +1,7 @@
 package client.scenes;
 
 
+import jakarta.ws.rs.BadRequestException;
 import commons.SubTask;
 import javafx.application.Platform;
 
@@ -17,10 +18,12 @@ import commons.Card;
 import commons.Listing;
 import jakarta.ws.rs.WebApplicationException;
 
+import javafx.fxml.FXML;
 import javafx.geometry.Bounds;
 import javafx.geometry.Insets;
 import javafx.geometry.Point2D;
 import javafx.geometry.Pos;
+import javafx.scene.Cursor;
 import javafx.scene.Node;
 import javafx.scene.Scene;
 import javafx.scene.control.*;
@@ -30,6 +33,7 @@ import javafx.scene.control.Button;
 import javafx.scene.control.TitledPane;
 import javafx.scene.paint.Color;
 import javafx.scene.text.Font;
+import javafx.scene.text.TextAlignment;
 import javafx.scene.text.Text;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
@@ -55,6 +59,7 @@ public class BoardOverviewController {
     private ServerUtils server;
     private ListController listController;
     private EventTarget target;
+    private boolean adminControl = false;
 
     // A map that will keep track of all dependencies between
     // the lists in the UI and the lists we have in the DB
@@ -142,7 +147,7 @@ public class BoardOverviewController {
         dialog.setHeaderText("Please enter a name for the card:");
         dialog.showAndWait().ifPresent(name -> {
 
-            if(!name.isEmpty()) {
+            if (!name.isEmpty()) {
                 VBox vBox = (VBox) addCardButton.getParent().getParent();
 
 
@@ -205,7 +210,7 @@ public class BoardOverviewController {
         dialog.setHeaderText("Please enter the new name for the card:");
         dialog.showAndWait().ifPresent(name -> {
 
-            if(!name.isEmpty()) {
+            if (!name.isEmpty()) {
                 cardButton.setText(name);
                 server.updateCard(currentCard.getCardId(), name);
             } else {
@@ -260,36 +265,41 @@ public class BoardOverviewController {
      * @throws IOException the exemption it might be caused
      */
     public void switchToInitialOverviewScene(javafx.event.ActionEvent actionEvent) throws IOException {
-        var initialOverview = FXML.load(InitialOverviewController.class, "client", "scenes", "InitialOverview.fxml");
-        initialOverview.getKey().setFileName(fileName);
-        initialOverview.getKey().refresh();
-        primaryStage = (Stage) ((Node) actionEvent.getSource()).getScene().getWindow();
-        overview = new Scene(initialOverview.getValue());
-        primaryStage.setScene(overview);
+        if (!adminControl) {
+            var initialOverview = FXML.load(InitialOverviewController.class, "client", "scenes", "InitialOverview.fxml");
+            initialOverview.getKey().setFileName(fileName);
+            initialOverview.getKey().refresh();
+            primaryStage = (Stage) ((Node) actionEvent.getSource()).getScene().getWindow();
+            overview = new Scene(initialOverview.getValue());
+            primaryStage.setScene(overview);
 
-        primaryStage.show();
+            primaryStage.show();
+        }
     }
 
     /**
      * Function that goes to the card details.
      *
      * @param actionEvent the action event on the button
-     * @param cardID the id of a card
-     * @param list the list of the card
+     * @param cardID      the id of a card
+     * @param list        the list of the card
      * @throws IOException the exception which might be caused
      */
     public void switchToCardScene(MouseEvent actionEvent, long cardID, Listing list) throws IOException {
-        var cardOverview = FXML.load(CardOverviewController.class, "client", "scenes", "CardOverview.fxml");
-        cardOverview.getKey().setCardId(cardID);
-        cardOverview.getKey().setFileName(fileName);
-        cardOverview.getKey().setBoard(board);
-        cardOverview.getKey().setList(list);
-        cardOverview.getKey().refresh();
-        primaryStage = (Stage) ((Node) actionEvent.getSource()).getScene().getWindow();
-        overview = new Scene(cardOverview.getValue());
-        primaryStage.setScene(overview);
-        primaryStage.show();
+        if (!adminControl) {
+            var cardOverview = FXML.load(CardOverviewController.class, "client", "scenes", "CardOverview.fxml");
+            cardOverview.getKey().setCardId(cardID);
+            cardOverview.getKey().setFileName(fileName);
+            cardOverview.getKey().setBoard(board);
+            cardOverview.getKey().setList(list);
+            cardOverview.getKey().refreshCardDetails();
+            primaryStage = (Stage) ((Node) actionEvent.getSource()).getScene().getWindow();
+            overview = new Scene(cardOverview.getValue());
+            primaryStage.setScene(overview);
+            primaryStage.show();
+        }
     }
+
 
     /**
      * This method handles dropping a hbox in another titledPane or within the same titledPane.
@@ -372,7 +382,6 @@ public class BoardOverviewController {
     }
 
 
-
     /**
      * addList method which accepts a Listing as a parameter.
      * <h5>NOTE: The IDs of the cards are stored within their user data.</h5>
@@ -396,7 +405,7 @@ public class BoardOverviewController {
         vBox.setPadding(new Insets(20, 0, 0, 0));
         vBox.setAlignment(Pos.TOP_CENTER);
         for (Card c : listing.getCards()) {
-            addCard(c,vBox,listing);
+            addCard(c, vBox, listing);
         }
         // add the "Add card" button below the cards
         HBox addCardButtonRow = new HBox();
@@ -423,8 +432,9 @@ public class BoardOverviewController {
 
     /**
      * Adds a card to the vBox List.
-     * @param c - the card we add
-     * @param vBox - the vBox which contains the list
+     *
+     * @param c       - the card we add
+     * @param vBox    - the vBox which contains the list
      * @param listing - the list the card is in
      */
     public void addCard (Card c, VBox vBox, Listing listing)
@@ -458,6 +468,7 @@ public class BoardOverviewController {
        // newCard.setPrefHeight(100);
         newCard.setUserData(c.getCardId());
         setupButton(newCard);
+        newCard.setCursor(Cursor.CLOSED_HAND);
         // make this card draggable
         newCard.setOnMousePressed(event -> {
             target = newCard.getParent(); // this is the hBox that needs to be dropped
@@ -468,7 +479,7 @@ public class BoardOverviewController {
         Button edit = new Button("\uD83D\uDD89");
         //edit.setOnAction(this::editCard); // an event happens when the button is clicked
         edit.setOnMousePressed(event -> {
-            if(event.getClickCount() == 2) {
+            if (event.getClickCount() == 2) {
                 try {
                     switchToCardScene(event, c.getCardId(), listing);
                 } catch (IOException e) {
@@ -494,8 +505,20 @@ public class BoardOverviewController {
      */
     public void refresh() {
         long id = board.getBoardId();
-        if (id != 0)
-            board = server.getBoardByID(id);
+        hBox.getChildren().clear();
+        if (id != 0) {
+            try {
+                board = server.getBoardByID(id);
+            } catch (BadRequestException e) {
+                Label noBoard = new Label("The board you are trying to access may have been deleted or does not exist.");
+                noBoard.setWrapText(true);
+                noBoard.setTextAlignment(TextAlignment.CENTER);
+                noBoard.setPrefWidth(500);
+                noBoard.setFont(new Font(20));
+                hBox.getChildren().add(noBoard);
+                return;
+            }
+        }
         listController.setBoard(board);
         String boardTitle = board.getTitle();
         Text boardText = new Text(boardTitle);
@@ -504,7 +527,6 @@ public class BoardOverviewController {
         renameBoardButton.setLayoutX(boardName.getLayoutX() + boardText.getLayoutBounds().getWidth() + 10.0);
         accessKey.setText("Access key: " + board.getAccessKey());
         List<Listing> listings = board.getLists();
-        hBox.getChildren().clear();
         map = new HashMap<>();
         for (Listing listing : listings)
             addListWithListing(listing);
@@ -535,40 +557,44 @@ public class BoardOverviewController {
      * @param actionEvent the event that triggered this method
      */
     public void leaveBoard(ActionEvent actionEvent) {
-        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
-        alert.setTitle("Confirmation Dialog");
-        alert.setHeaderText("Are you sure you want to leave this board?");
-        alert.setContentText("You will not be able to access this board again.");
-        Optional<ButtonType> result = alert.showAndWait();
-        if (result.isPresent() && result.get() == ButtonType.OK) {
-            long id = board.getBoardId();
-            File file = new File(fileName);
-            StringBuilder content = new StringBuilder();
-            try {
-                Scanner scanner = new Scanner(file);
-                while (scanner.hasNextLine()) {
-                    String line = scanner.nextLine();
-                    if (!line.startsWith(Long.toString(id))) {
-                        content.append(line).append("\n");
+        if (!adminControl) {
+            Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+            alert.setTitle("Confirmation Dialog");
+            alert.setHeaderText("Are you sure you want to leave this board?");
+            alert.setContentText("You will not be able to access this board again.");
+            Optional<ButtonType> result = alert.showAndWait();
+            if (result.isPresent() && result.get() == ButtonType.OK) {
+                long id = board.getBoardId();
+                File file = new File(fileName);
+                String content = "";
+                try {
+                    Scanner scanner = new Scanner(file);
+                    while (scanner.hasNextLine()) {
+                        String line = scanner.nextLine();
+                        if (!line.startsWith(Long.toString(id))) {
+                            content = content + line + "\n";
+                        }
                     }
+                    FileOutputStream outputStream = new FileOutputStream(file);
+                    outputStream.write(content.getBytes());
+                    outputStream.flush();
+                    outputStream.close();
+                    switchToInitialOverviewScene(actionEvent);
+                } catch (IOException e) {
+                    e.printStackTrace();
                 }
-                FileOutputStream outputStream = new FileOutputStream(file);
-                outputStream.write(content.toString().getBytes());
-                outputStream.flush();
-                outputStream.close();
-                switchToInitialOverviewScene(actionEvent);
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
 
+
+            }
         }
     }
 
     /**
      * Sets up the delete list button.
+     *
      * @param button the button to set up
      */
-    private void setupDeleteListButton(Button button){
+    private void setupDeleteListButton(Button button) {
         button.setBorder(new Border(new BorderStroke(Color.RED, BorderStrokeStyle.SOLID, new CornerRadii(5), BorderWidths.DEFAULT)));
         button.setStyle("-fx-background-color: transparent;");
         button.setTextFill(Color.RED);
@@ -584,9 +610,10 @@ public class BoardOverviewController {
 
     /**
      * Sets up the add card button.
+     *
      * @param button the button to set up
      */
-    private void setupAddCardButton(Button button){
+    private void setupAddCardButton(Button button) {
         button.setBorder(new Border(new BorderStroke(Color.GREEN, BorderStrokeStyle.SOLID, new CornerRadii(5), BorderWidths.DEFAULT)));
         button.setStyle("-fx-background-color: transparent;");
         button.setTextFill(Color.GREEN);
@@ -599,12 +626,14 @@ public class BoardOverviewController {
             button.setTextFill(Color.GREEN);
         });
     }
+
     /**
      * Sets up the buttons contained in the lists.
+     *
      * @param button the button to set up
      */
-    private void setupButton(Button button){
-        HBox.setHgrow(button,Priority.ALWAYS);
+    private void setupButton(Button button) {
+        HBox.setHgrow(button, Priority.ALWAYS);
         button.setStyle("-fx-background-color: transparent");
         button.setMaxWidth(Double.MAX_VALUE);
         button.setMinWidth(Button.USE_PREF_SIZE);
@@ -612,6 +641,13 @@ public class BoardOverviewController {
         button.setOnMouseExited(event -> button.setStyle("-fx-background-color: transparent;"));
     }
 
+    /**
+     * Setter for the admin control value, which determines whether the app was opened in admin control mode.
+     * @param adminControl the value to set
+     */
+    public void setAdminControl(boolean adminControl) {
+        this.adminControl = adminControl;
+    }
     /**
      * Copies the board's access key to the clipboard.
      * @param mouseEvent the event that triggered this method
